+++ photons_tile_paint/gameoflife/options.py
@@ -1,4 +1,4 @@
-from photons_tile_paint.options import BackgroundOption
+from photons_tile_paint.options import AnimationOptions
 
 from photons_themes.theme import ThemeColor as Color
 
@@ -6,8 +6,7 @@ from input_algorithms.dictobj import dictobj
 from input_algorithms import spec_base as sb
 import random
 
-class TileGameOfLifeOptions(dictobj.Spec):
-    background = dictobj.Field(BackgroundOption.FieldSpec())
+class TileGameOfLifeOptions(AnimationOptions):
     user_coords = dictobj.Field(sb.boolean, default=True)
     num_iterations = dictobj.Field(sb.integer_spec, default=-1)
     new_color_style = dictobj.Field(sb.string_choice_spec(["random", "average"]), default="average")