+++ photons_tile_paint/marquee/options.py
@@ -1,12 +1,11 @@
-from photons_tile_paint.options import BackgroundOption, ColorOption
+from photons_tile_paint.options import AnimationOptions, ColorOption
 from photons_tile_paint.marquee.animation import MarqueeDirection
 from photons_protocol.types import enum_spec
 
 from input_algorithms.dictobj import dictobj
 from input_algorithms import spec_base as sb
 
-class TileMarqueeOptions(dictobj.Spec):
-    background = dictobj.Field(BackgroundOption.FieldSpec())
+class TileMarqueeOptions(AnimationOptions):
     text_color = dictobj.Field(ColorOption(200, 0.24, 0.5, 3500))
     text = dictobj.Field(sb.string_spec, default="LIFX is awesome!")
     user_coords = dictobj.Field(sb.boolean, default=False)