+++ photons_tile_paint/pacman/animation.py
@@ -8,8 +8,6 @@ class TilePacmanAnimation(Animation):
 
     def setup(self):
         self.iteration = 0
-        if self.options.user_coords:
-            self.coords = None
 
     def next_state(self, prev_state, coords):
         if prev_state is None: