+++ photons_tile_paint/pacman/options.py
@@ -1,10 +1,9 @@
-from photons_tile_paint.options import BackgroundOption
+from photons_tile_paint.options import AnimationOptions
 
 from input_algorithms.dictobj import dictobj
 from input_algorithms import spec_base as sb
 
-class TilePacmanOptions(dictobj.Spec):
-    background = dictobj.Field(BackgroundOption.FieldSpec())
+class TilePacmanOptions(AnimationOptions):
     user_coords = dictobj.Field(sb.boolean, default=False)
     num_iterations = dictobj.Field(sb.integer_spec, default=-1)